third_party/blink/web_tests/external/wpt/html/dom/elements/global-attributes/the-anchor-attribute-xml.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] the-anchor-attribute-xml
  assert_equals: Setting the anchor attribute in XML should not set the anchorElement IDL. expected null but got Element node <div id="target">target</div>
Harness: the test ran to completion.
